+++ fixtures/services-node-basic/CHANGELOG.md
@@ -1,5 +1,11 @@
 # services-node-basic-fixture
 
+## 0.0.14
+
+### Patch Changes
+
+- theokit@0.9.7
+
 ## 0.0.13
 
 ### Patch Changes